postino_notification_service-develop/prod.requirements.txt
fastapi[standard]==0.115.12
pydantic==2.11.3
pydantic-settings==2.9.1
httpx==0.28.1
dependency-injector==4.46.0
python-dotenv==1.1.0
orjson==3.10.18
uvicorn==0.34.2
phonenumbers==9.0.4
pydantic-extra-types==2.10.3
gunicorn==23.0.0